requirements.txt
torch
numpy
pyyaml
pillow
tqdm
